add CovarianceMixin
